+++ Auxiliary/DirectXTexPNG.cpp
@@ -327,7 +327,7 @@ HRESULT DirectX::GetMetadataFromPNGFile(
     catch (const std::system_error& ec)
     {
 #ifdef _WIN32
-        return HRESULT_FROM_WIN32(ec.code().value());
+        return HRESULT_FROM_WIN32(static_cast<unsigned long>(ec.code().value()));
 #else
         return (ec.code().value() == ENOENT) ? HRESULT_ERROR_FILE_NOT_FOUND : E_FAIL;
 #endif
@@ -368,7 +368,7 @@ HRESULT DirectX::LoadFromPNGFile(
     {
         image.Release();
 #ifdef _WIN32
-        return HRESULT_FROM_WIN32(ec.code().value());
+        return HRESULT_FROM_WIN32(static_cast<unsigned long>(ec.code().value()));
 #else
         return (ec.code().value() == ENOENT) ? HRESULT_ERROR_FILE_NOT_FOUND : E_FAIL;
 #endif
@@ -402,7 +402,7 @@ HRESULT DirectX::SaveToPNGFile(
     catch (const std::system_error& ec)
     {
 #ifdef _WIN32
-        return HRESULT_FROM_WIN32(ec.code().value());
+        return HRESULT_FROM_WIN32(static_cast<unsigned long>(ec.code().value()));
 #else
         return (ec.code().value() == ENOENT) ? HRESULT_ERROR_FILE_NOT_FOUND : E_FAIL;
 #endif
